+++ custom_components/wattpilot/select.py
@@ -29,6 +29,7 @@ if TYPE_CHECKING:
 
 _LOGGER: Final = logging.getLogger(__name__)
 PLATFORM = "select"
+PARALLEL_UPDATES = 1
 
 
 async def async_setup_entry(