+++ custom_components/wattpilot/switch.py
@@ -31,6 +31,7 @@ if TYPE_CHECKING:
 
 _LOGGER: Final = logging.getLogger(__name__)
 PLATFORM = "switch"
+PARALLEL_UPDATES = 1
 
 
 async def async_setup_entry(