+++ custom_components/wattpilot/update.py
@@ -38,6 +38,7 @@ if TYPE_CHECKING:
 
 _LOGGER: Final = logging.getLogger(__name__)
 PLATFORM = "update"
+PARALLEL_UPDATES = 1
 
 
 async def async_setup_entry(